+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0).
 
 ## Unreleased
 
-Nothing.
+### Added
+
+- Print elapsed time for each command.
 
 ## [2.0.2](https://github.com/trallnag/exec-cmds-defer-errors/compare/v2.0.1...v2.0.2) / 2025-03-01